Normalize song data for searching and browsing.
Add ArtistNorm, TitleNorm, and AlbumNorm columns containing
normalized (decomposed, accent-stripped, lowercased, and
trimmed) strings to the Songs database table. This makes it
possible to perform local searches using easier-to-type
strings and also improves aggregation when browsing.

As part of the same schema change, store song lengths as
floats rather than ints.

Fixes #21.

+++ nup/src/main/java/MediaSessionManager.kt
@@ -54,7 +54,7 @@ class MediaSessionManager constructor(context: Context, callback: MediaSessionCo
         if (song.disc > 0) {
             builder.putLong(MediaMetadataCompat.METADATA_KEY_DISC_NUMBER, song.disc.toLong())
         }
-        builder.putLong(MediaMetadataCompat.METADATA_KEY_DURATION, song.lengthSec * 1000L)
+        builder.putLong(MediaMetadataCompat.METADATA_KEY_DURATION, (song.lengthSec * 1000).toLong())
 
         val bitmap = song.coverBitmap
         if (bitmap != null) {
